+++ fmm/src/analysis/c_calling_convention/context.rs
@@ -1,0 +1,13 @@
+pub struct Context {
+    word_bytes: usize,
+}
+
+impl Context {
+    pub fn new(word_bytes: usize) -> Self {
+        Self { word_bytes }
+    }
+
+    pub fn word_bytes(&self) -> usize {
+        self.word_bytes
+    }
+}
