+++ fmm/src/analysis/c_calling_convention/utilities.rs
@@ -1,0 +1,20 @@
+use super::context::Context;
+use crate::{analysis::type_size, types::Type};
+
+// The name, "memory class" comes from the C ABI on System V.
+pub fn is_memory_class(context: &Context, type_: &Type) -> bool {
+    match type_ {
+        Type::Record(record) => {
+            type_size::calculate_size(type_, context.word_bytes()) > 2 * context.word_bytes()
+                || record
+                    .fields()
+                    .iter()
+                    .any(|type_| is_memory_class(context, type_))
+        }
+        Type::Union(union) => union
+            .members()
+            .iter()
+            .any(|type_| is_memory_class(context, type_)),
+        Type::Function(_) | Type::Pointer(_) | Type::Primitive(_) => false,
+    }
+}
